FASTAPI-MoM/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic>=2,<3
asyncpg

